+++ FreeRTOS/Demo/Posix_GCC/code_coverage_additions.c
