+++ FreeRTOS/Demo/Posix_GCC/trcConfig.h
@@ -113,7 +113,7 @@ extern "C" {
  * TRC_FREERTOS_VERSION_9_0_2					If using FreeRTOS v9.0.2
  * TRC_FREERTOS_VERSION_10_0_0					If using FreeRTOS v10.0.0 or later
  *****************************************************************************/
-#define TRC_CFG_FREERTOS_VERSION TRC_FREERTOS_VERSION_10_0_0
+#define TRC_CFG_FREERTOS_VERSION TRC_FREERTOS_VERSION_10_4_0
 
 /*******************************************************************************
  * TRC_CFG_SCHEDULING_ONLY
